refactor: JDTTreeBuilder, split into multiple classes with one static Helper
