+++ src/main/java/spoon/support/compiler/jdt/ASTPair.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright (C) 2006-2016 INRIA and contributors
+ * Spoon - http://spoon.gforge.inria.fr/
+ *
+ * This software is governed by the CeCILL-C License under French law and
+ * abiding by the rules of distribution of free software. You can use, modify
+ * and/or redistribute the software under the terms of the CeCILL-C license as
+ * circulated by CEA, CNRS and INRIA at http://www.cecill.info.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the CeCILL-C License for more details.
+ *
+ * The fact that you are presently reading this means that you have had
+ * knowledge of the CeCILL-C license and that you accept its terms.
+ */
+package spoon.support.compiler.jdt;
+
+import org.eclipse.jdt.internal.compiler.ast.ASTNode;
+import spoon.reflect.declaration.CtElement;
+
+public class ASTPair {
+	public CtElement element;
+
+	public ASTNode node;
+
+	public ASTPair(CtElement element, ASTNode node) {
+		super();
+		this.element = element;
+		this.node = node;
+	}
+
+	@Override
+	public String toString() {
+		return element.getClass().getSimpleName() + "-" + node.getClass().getSimpleName();
+	}
+}
